cusatconnect-backend/requirements.txt
Django==2.2.1
django-cors-headers==3.0.2
django-expiring-token==1.0.2
djangorestframework==3.9.4
mysqlclient==1.4.4